LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/pointerevents/pointerevent_after_target_removed_from_slot_mouse-expected.txt

done

FAIL Pointer events from mouse received before/after slot removal at pointerdown assert_equals: events received after removal expected "pointerover@parent,pointerover@host,pointerup@parent,pointerup@host,pointerdown@parent,pointerdown@host,pointerup@parent,pointerup@host,pointerout@parent,pointerout@host,pointerleave@parent,pointerleave@host" but got "pointerup@parent,pointerup@host,pointerout@child,pointerleave@child,pointerover@host,pointerover@parent,pointerenter@parent,pointerdown@parent,pointerdown@host,pointerup@parent,pointerup@host,pointerout@parent,pointerout@host,pointerleave@parent,pointerleave@host"
FAIL Pointer events from mouse received before/after slotted-child removal at pointerdown assert_equals: events received after removal expected "pointerleave@slot,pointerover@parent,pointerover@host,pointerup@parent,pointerup@host,pointerdown@parent,pointerdown@host,pointerup@parent,pointerup@host,pointerout@parent,pointerout@host,pointerleave@parent,pointerleave@host" but got "pointerup@parent,pointerup@host,pointerover@parent,pointerover@host,pointerenter@host,pointerenter@parent,pointerdown@parent,pointerdown@host,pointerup@parent,pointerup@host,pointerout@parent,pointerout@host,pointerleave@parent,pointerleave@host"
FAIL Pointer events from mouse received before/after slot removal at pointerup assert_equals: events received after removal expected "pointerover@parent,pointerover@host,pointerdown@parent,pointerdown@host,pointerup@parent,pointerup@host,pointerout@parent,pointerout@host,pointerleave@parent,pointerleave@host" but got "pointerout@child,pointerleave@child,pointerover@parent,pointerover@host,pointerenter@parent,pointerdown@parent,pointerdown@host,pointerup@parent,pointerup@host,pointerout@parent,pointerout@host,pointerleave@parent,pointerleave@host"
FAIL Pointer events from mouse received before/after slotted-child removal at pointerup assert_equals: events received after removal expected "pointerleave@slot,pointerover@parent,pointerover@host,pointerdown@parent,pointerdown@host,pointerup@parent,pointerup@host,pointerout@parent,pointerout@host,pointerleave@parent,pointerleave@host" but got "pointerover@parent,pointerover@host,pointerenter@host,pointerenter@parent,pointerdown@parent,pointerdown@host,pointerup@parent,pointerup@host,pointerout@parent,pointerout@host,pointerleave@parent,pointerleave@host"
